+++ inst/doc/v5_ASOIAF.R
@@ -2,7 +2,7 @@
 library(BGmisc)
 library(tidyverse)
 library(ggpedigree)
-data(ASOIAF)
+data(ASOIAF, package = "ggpedigree")
 
 ## -----------------------------------------------------------------------------
 head(ASOIAF)
@@ -63,7 +63,6 @@ jon_dany_row <- df_links %>%
 jon_dany_row
 
 ## -----------------------------------------------------------------------------
-
 rhaenyra_id <- df_got %>%
   filter(name == "Rhaenyra Targaryen") %>%
   pull(ID)
@@ -71,15 +70,15 @@ daemon_id <- df_got %>%
   filter(name == "Daemon Targaryen") %>%
   pull(ID)
 
-rhaenyra_daemon_row <- df_links %>% 
+rhaenyra_daemon_row <- df_links %>%
   filter(ID1 == rhaenyra_id | ID2 == rhaenyra_id) %>%
-  filter(ID1 %in% daemon_id | ID2 %in% daemon_id) %>% # round to nearest 4th decimal
+  filter(ID1 %in% daemon_id | ID2 %in% daemon_id) %>% # round to 4th decimal
   mutate(across(c(addRel, mitRel, cnuRel), ~ round(.x, 4)))
 
 rhaenyra_daemon_row
 
 ## -----------------------------------------------------------------------------
-df_repaired <- checkParentIDs(df_got,
+df_repaired <- checkParentIDs(df_got, # %>% filter(famID == 1),
   addphantoms = TRUE,
   repair = TRUE,
   parentswithoutrow = FALSE,
@@ -87,20 +86,19 @@ df_repaired <- checkParentIDs(df_got,
 ) %>% mutate(
   # famID = 1,
   affected = case_when(
-    ID %in% c(jon_id, dany_id, "365") ~ T,
-    TRUE ~ F
+    ID %in% c(jon_id, dany_id, 339) ~ TRUE,
+    TRUE ~ FALSE
   )
 )
 
 ## -----------------------------------------------------------------------------
-
 checkIDs <- checkIDs(df_repaired, verbose = TRUE)
 
-#checkIDs
+# checkIDs
 
 ## -----------------------------------------------------------------------------
 # Check for unique IDs and parent-child relationships
-checkPedigreeNetwork<- checkPedigreeNetwork(df_repaired,
+checkPedigreeNetwork <- checkPedigreeNetwork(df_repaired,
   personID = "ID",
   momID = "momID",
   dadID = "dadID",
@@ -110,10 +108,13 @@ checkPedigreeNetwork<- checkPedigreeNetwork(df_repaired,
 checkPedigreeNetwork
 
 ## ----eval=FALSE, fig.height=8, fig.width=10, message=FALSE, warning=FALSE, include=FALSE----
-# plotPedigree(df_repaired %>% mutate(
-#   famID = 1
-# ), affected = df_repaired$affected,
-# verbose = FALSE)
+# plotPedigree(
+#   df_repaired %>% mutate(
+#     famID = 1
+#   ),
+#   affected = df_repaired$affected,
+#   verbose = FALSE
+# )
 
 ## ----message=FALSE, warning=FALSE, fig.width=10, fig.height=8-----------------
 library(ggpedigree)
@@ -126,34 +127,41 @@ plt <- ggpedigree(df_repaired_renamed,
   personID = "personID",
   interactive = FALSE,
   config = list(
-   overlay_include = TRUE,
+    overlay_include = TRUE,
     point_size = .75,
     code_male = "M",
-    ped_width = 14,
+    ped_width = 17,
     label_nudge_y = -.25,
     include_labels = TRUE,
     label_method = "geom_text",
-    #segment_self_color = "purple",
+    # segment_self_color = "purple",
     sex_color_include = FALSE,
-    focal_fill_personID = 339,#353,
+    focal_fill_personID = 353, # 339, # 353,
     focal_fill_include = TRUE,
-    tooltip_columns = c( "personID","name", "focal_fill"),
+    tooltip_columns = c("personID", "name", "focal_fill"),
     focal_fill_force_zero = TRUE,
     focal_fill_mid_color = "orange",
     focal_fill_low_color = "#9F2A63FF",
-    focal_fill_legend_title = "Relatedness to \nRhaenyra Targaryen",
+    focal_fill_legend_title = "Relatedness to \nAegon Targaryen",
     focal_fill_na_value = "black",
     value_rounding_digits = 4
-))
+  )
+)
 
 plt
 
 # reduce file size for CRAN
-#if (interactive()) {
-  # If running interactively, use plotly::partial_bundle
-  # to reduce file size for CRAN
+# if (interactive()) {
+# If running interactively, use plotly::partial_bundle
+# to reduce file size for CRAN
 #  plotly::partial_bundle(plt)
-#} else {
+# } else {
 #  plotly::partial_bundle(plt, local = TRUE)
-#}
+# }
+
+## ----eval=FALSE, include=FALSE------------------------------------------------
+# df_repaired %>%
+#   filter(!is.na(name)) %>%
+#   arrange(ID) %>%
+#   knitr::kable(caption = "Key Characters in ASOIAF Pedigree")