+++ tests/testthat/test-checkParents.R
@@ -14,9 +14,9 @@ test_that("checkParentIDs identifies parent coding correctly in potter dataset",
 
 # Test Case 2: Validate sex coding without repair
 test_that("checksif single parents found correctly in ASOIAF dataset", {
-  data(ASOIAF)
+  data(ASOIAF, package = "ggpedigree")
   df_asoiaf <- ASOIAF
-  results <- checkParentIDs(df_asoiaf, verbose = FALSE, repair = FALSE)
+  results <- checkParentIDs(df_asoiaf, verbose = FALSE, repair = FALSE, personID = "id")
   expect_true(results$single_parents)
   single_dads <- length(df_asoiaf$id[!is.na(df_asoiaf$dadID) & is.na(df_asoiaf$momID)])
   single_moms <- length(df_asoiaf$id[is.na(df_asoiaf$dadID) & !is.na(df_asoiaf$momID)])
@@ -27,8 +27,9 @@ test_that("checksif single parents found correctly in ASOIAF dataset", {
 })
 
 test_that("verbose checks", {
-  data(ASOIAF)
+  data(ASOIAF, package = "ggpedigree")
   df_asoiaf <- ASOIAF
   expect_message(checkParentIDs(df_asoiaf, verbose = TRUE, repair = TRUE))
+  expect_message(repairParentIDs(df_asoiaf, verbose = TRUE))
   expect_message(checkParentIDs(df_asoiaf, verbose = TRUE, parentswithoutrow = TRUE))
 })